+++ src/job.h
@@ -27,7 +27,7 @@ struct JobTable {
   struct detail;
   std::unique_ptr<detail> imp;
 
-  JobTable(Database *db, int max_jobs, bool verbose, bool quiet, bool check);
+  JobTable(Database *db, double max_jobs, bool verbose, bool quiet, bool check);
   ~JobTable();
 
   // Wait for a job to complete; false -> no more active jobs
